Changing TemplateSample's gray value back to u8 to save memory, now that we're trying a different kind of intensity normalization (that runs completely as a preprocessor to tracking and provides already-normalized u8 images).

+++ coretech/vision/robot/src/lucasKanade_SampledPlanar6dof.cpp
@@ -578,7 +578,7 @@ namespace Anki
 #if USE_INTENSITY_NORMALIZATION
               curTemplateSample.grayvalue   = (static_cast<f32>(pGrayscaleVector[curIndex]) - curMean) * curSigmaInv;
 #else
-              curTemplateSample.grayvalue   = static_cast<f32>(pGrayscaleVector[curIndex]);
+              curTemplateSample.grayvalue   = pGrayscaleVector[curIndex];
 #endif
               
               // Everything below here is about filling in the Jacobian info
